requirements.txt
# Framework
fastapi==0.111.0
uvicorn
pydantic
## jwt
python-jose>=3.3.0
# Secret
cryptography==42.0.5
# DB
SQLAlchemy==2.0.30
sqlalchemy-doris[pymysql]==0.2.2
redis==3.5.2
confluent-kafka==2.4.0
# Other
requests==2.31.0
loguru==0.7.2
